# infra/cdk/openapi.yaml
openapi: 3.0.3
info:
  title: DeepCopy API
  version: 1.0.0
  description: |
    API for submitting AI processing jobs and retrieving their status and results.

    Authentication (OAuth2 client credentials via Amazon Cognito):

    ```bash
    # Set endpoints (from CDK outputs)
    API_URL="https://o5egokjpsl.execute-api.eu-west-1.amazonaws.com/prod/"
    TOKEN_ENDPOINT="https://deepcopy-613663743323-eu-west-1.auth.eu-west-1.amazoncognito.com/oauth2/token"

    # Your app client credentials (retrieve secret from Cognito console)
    CLIENT_ID="<your-client-id>"
    CLIENT_SECRET="<your-client-secret>"

    # Request access token with read+write scopes
    ACCESS_TOKEN=$(curl -s -u "$CLIENT_ID:$CLIENT_SECRET" \
      -d "grant_type=client_credentials&scope=https://deep-copy.api/read https://deep-copy.api/write" \
      "$TOKEN_ENDPOINT" | jq -r .access_token)

    echo "Token acquired: ${#ACCESS_TOKEN} chars"

    # Submit a v2 job
    curl -sS -X POST \
      -H "Authorization: Bearer $ACCESS_TOKEN" \
      -H "content-type: application/json" \
      -d '{
        "sales_page_url": "https://try.meritrelief.com/",
        "project_name": "meritrelief",
        "advertorial_type": "Listicle",
        "gender": "Female"
      }' \
      "${API_URL}v2/jobs"

    # Get v2 job status (replace with real jobId)
    curl -sS -H "Authorization: Bearer $ACCESS_TOKEN" "${API_URL}v2/jobs/YOUR_JOB_ID"

    # Get v2 job result JSON (404 until available)
...
    ```
servers:
...
security:
  - cognitoOAuth2: [ ]
paths:
...
  /jobs/{id}/result:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    options:
      summary: CORS preflight
      responses:
        '204':
          description: No Content
          headers:
            Access-Control-Allow-Origin: { $ref: '#/components/headers/AccessControlAllowOrigin' }
            Access-Control-Allow-Headers: { $ref: '#/components/headers/AccessControlAllowHeaders' }
            Access-Control-Allow-Methods: { $ref: '#/components/headers/AccessControlAllowMethods' }
    get:
      summary: Get job result
      operationId: getJobResult
      security:
        - cognitoOAuth2:
            - https://deep-copy.api/read
      responses:
        '200':
          description: Job result JSON or text
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResult'
            text/plain:
              schema:
                type: string
        '404':
          description: Result not available
          content:
            text/plain:
              schema:
                type: string
                example: "Result not available"
        '500':
          description: Server error
          content:
            text/plain:
              schema:
                type: string
      x-codeSamples:
        - lang: bash
          label: curl (result)
          source: |-
            API_URL="https://o5egokjpsl.execute-api.eu-west-1.amazonaws.com/prod/"
            curl -sS -H "Authorization: Bearer $ACCESS_TOKEN" "${API_URL}jobs/YOUR_JOB_ID/result"
...
components:
  securitySchemes:
    cognitoOAuth2:
      type: oauth2
      description: |
        OAuth2 client credentials against Amazon Cognito User Pool domain.
        Use the token endpoint output by the stack to obtain an access token.
      flows:
        clientCredentials:
          tokenUrl: https://deepcopy-613663743323-eu-west-1.auth.eu-west-1.amazoncognito.com/oauth2/token
          scopes:
            https://deep-copy.api/read: Read jobs and results
            https://deep-copy.api/write: Submit jobs
  schemas:
...
    JobResult:
      type: object
      description: Structured JSON produced by the pipeline including rewritten advertorial content.
      additionalProperties: false
      properties:
        project_name:
          type: string
        timestamp_iso:
          type: string
          format: date-time
        job_id:
          type: string
        api_version:
          type: string
          description: API version (v1)
        results:
          type: object
          additionalProperties: false
          properties:
            research_page_analysis:
              type: string
            doc1_analysis:
              type: string
            doc2_analysis:
              type: string
            deep_research_prompt:
              type: string
            deep_research_output:
              type: string
            avatar_sheet:
              $ref: '#/components/schemas/AvatarList'
            identified_avatars:
              $ref: '#/components/schemas/IdentifiedAvatarList'
            offer_brief:
              $ref: '#/components/schemas/OfferBrief'
            marketing_philosophy_analysis:
              type: string
            necessary_beliefs:
              type: string
            summary:
              type: string
            swipe_results:
              type: array
              items:
                $ref: '#/components/schemas/SwipeResult'
            marketing_angles:
              type: array
              items:
                $ref: '#/components/schemas/Angle'
            customer_avatars:
              type: array
              description: Simplified list of identified avatars
              items:
                $ref: '#/components/schemas/IdentifiedAvatar'
      required: [project_name, timestamp_iso, results, job_id]
    SwipeResult:
      type: object
      description: Result item for a specific marketing angle containing structured advertorial content.
      required: [angle, content]
      properties:
        angle:
          type: string
          description: Marketing angle used to generate the content.
        content:
          description: Structured content; either a Listicle or an Avertorial.
          oneOf:
            - $ref: '#/components/schemas/Listicle'
            - $ref: '#/components/schemas/Advertorial'
...
    Angle:
      type: object
      required: [angle, title, target_age_range, target_audience, pain_points, desires, common_objections, failed_alternatives, copy_approach]
      properties:
        angle:
          type: string
          description: "The marketing angle for the product"
        title:
          type: string
          description: "The title of the marketing angle"
        target_age_range:
          type: string
          description: "The target age range for the marketing angle. Note it like: 45-55, 60-70, etc."
        target_audience:
          type: string
          description: "The target audience for the marketing angle. Note it like: Competitive athletes seeking reliability, Skilled professionals seeking productivity, etc."
        pain_points:
          type: array
          description: "The pain points for the marketing angle"
          items:
            type: string
        desires:
          type: array
          description: "The desires for the marketing angle"
          items:
            type: string
        common_objections:
          type: array
          description: "The common objections for the marketing angle"
          items:
            type: string
        failed_alternatives:
          type: array
          description: "The failed alternatives for the marketing angle"
          items:
            type: string
        copy_approach:
          type: array
          description: "The copy approaches for the marketing angle"
          items:
            type: string
    # ===== New Avatar Enums =====
...
    Avatar:
      type: object
      description: "Complete Marketing Avatar sheet for product–market fit & messaging research"
      required: [id, short_description, age, gender, problem_urgency, purchasing_power, saturation_level, audience_size, overall_score, overview, demographics, problem_experience, pain_desire, failed_solutions, objections_buying, raw_language, advertising_platforms]
      properties:
        id:
          type: string
          format: uuid
          description: "Unique identifier for this avatar"
        short_description:
          type: string
          description: "Short description of the avatar"
          maxLength: 40
        age:
          type: string
          description: "Age range of the avatar"
          maxLength: 20
        gender:
          type: string
          enum: [Male, Female, Both]
          description: "Gender of the avatar"
        # Section 0: Avatar Evaluation Dimensions (1-5 scores)
        problem_urgency:
          type: integer
          minimum: 1
          maximum: 5
          description: "How desperately this avatar needs a solution (1-5). 5=Critical, 4=High, 3=Moderate, 2=Low, 1=Minimal"
        purchasing_power:
          type: integer
          minimum: 1
          maximum: 5
          description: "Can this avatar afford the solution (1-5). 5=Premium, 4=Comfortable, 3=Moderate, 2=Stretched, 1=Limited"
        saturation_level:
          type: integer
          minimum: 1
          maximum: 5
          description: "How crowded is the market for this avatar (1-5). 5=Blue Ocean, 4=Low, 3=Moderate, 2=High, 1=Saturated"
        audience_size:
          type: integer
          minimum: 1
          maximum: 5
          description: "How large is this avatar segment (1-5). 5=Massive (40%+), 4=Large (25-40%), 3=Medium (15-25%), 2=Small (5-15%), 1=Tiny (<5%)"
        overall_score:
          type: integer
          minimum: 1
          maximum: 5
          description: "Overall score of the avatar (1-5) based on problem urgency, purchasing power, saturation level, audience size"
        overview:
          $ref: '#/components/schemas/AvatarOverviewSection'
          description: "Section 1: Avatar Overview"
        demographics:
          $ref: '#/components/schemas/AvatarDemographicsSection'
          description: "Section 2: Demographics & Identity"
        problem_experience:
          $ref: '#/components/schemas/AvatarProblemExperienceSection'
          description: "Section 3: Problem Experience"
        pain_desire:
          $ref: '#/components/schemas/AvatarPainDesireSection'
          description: "Section 4: Pain & Desire"
        failed_solutions:
          $ref: '#/components/schemas/AvatarFailedSolutionsSection'
          description: "Section 5: Failed Solutions"
        objections_buying:
          $ref: '#/components/schemas/AvatarObjectionsBuyingSection'
          description: "Section 6: Objections & Buying"
        raw_language:
          $ref: '#/components/schemas/AvatarRawLanguageSection'
          description: "Section 7: Raw Language"
        advertising_platforms:
          type: string
          description: "List the advertising platforms this avatar primarily uses"
...
    IdentifiedAvatar:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
          description: Name of the avatar (e.g., 'Busy Mom')
        description:
          type: string
          description: Brief description of who they are and why they are a good fit
    IdentifiedAvatarList:
      type: object
      required: [avatars]
      properties:
        avatars:
          type: array
          items:
            $ref: '#/components/schemas/IdentifiedAvatar'
          description: List of potential customer avatars
    AvatarList:
      type: object
      required: [avatars]
      properties:
        avatars:
          type: array
          items:
            $ref: '#/components/schemas/Avatar'
          description: List of identified customer avatars
...
    MarketSnapshot:
      type: object
      description: "Section 1: Market Snapshot - always visible summary"
      properties:
        sophistication:
          $ref: '#/components/schemas/StageOfSophistication'
        awareness:
          $ref: '#/components/schemas/AwarenessLevel'
        awareness_description:
          type: string
          nullable: true
          description: Brief description of awareness state
        consciousness:
          $ref: '#/components/schemas/ConsciousnessLevel'
        consciousness_description:
          type: string
          nullable: true
          description: Brief description of consciousness state
        market_temperature:
          $ref: '#/components/schemas/MarketTemperature'
        market_temperature_description:
          type: string
          nullable: true
          description: Explanation of market temperature
    BigIdeaSection:
      type: object
      description: "Section 2: The Big Idea and mechanisms"
      properties:
        big_idea:
          type: string
          nullable: true
          description: Single, transformative central idea that unifies the campaign
        problem_mechanism_ump:
          type: string
          nullable: true
          description: "Unique Mechanism of the Problem: Why nothing has worked"
        solution_mechanism_ums:
          type: string
          nullable: true
          description: "Unique Mechanism of the Solution: Why THIS works"
        metaphors:
          type: array
          description: Images/analogies to make the big idea and mechanism vivid
          items:
            type: string
        guru:
          type: string
          nullable: true
          description: Authority or spokesperson (real or archetype) associated with the offer
        discovery_story:
          type: string
          nullable: true
          description: "Origin narrative: aha moment, failed attempts, breakthrough mechanism"
...
    ObjectionsSection:
      type: object
      description: "Section 8: Objections handling"
      properties:
        objections:
          type: array
          items:
            $ref: '#/components/schemas/ObjectionItem'
        hidden_objection:
          type: string
          nullable: true
        hidden_objection_counter:
          type: string
          nullable: true
...
    ProductInfo:
      type: object
      properties:
        name:
          type: string
          nullable: true
          description: Working name of the product/offer
        description:
          type: string
          nullable: true
          description: What it is, who it's for, core outcomes/benefits
        details:
          type: string
          nullable: true
          description: Format, modules, deliverables, bonuses, guarantees, price points, terms
        format:
          type: string
          nullable: true
          description: Product format
        price:
          type: string
          nullable: true
          description: One-time price
        subscription_price:
          type: string
          nullable: true
          description: Subscription/recurring price if applicable
        guarantee:
          type: string
          nullable: true
          description: Guarantee offered
        shipping:
          type: string
          nullable: true
          description: Shipping details
        key_differentiator:
          type: string
          nullable: true
          description: The primary thing that sets this product apart from competitors
        compliance_notes:
          type: array
          description: Legal/compliance requirements
          items:
            type: string
    OfferBrief:
      type: object
      description: Structured brief to guide AI when generating copy, concepts, and funnel assets
      properties:
        market_snapshot:
          $ref: '#/components/schemas/MarketSnapshot'
        big_idea_section:
          $ref: '#/components/schemas/BigIdeaSection'
        product:
          $ref: '#/components/schemas/ProductInfo'
        pain_desire:
          $ref: '#/components/schemas/PainDesireSection'
        failed_solutions:
          $ref: '#/components/schemas/FailedSolutionsSection'
        competitor_landscape:
          $ref: '#/components/schemas/CompetitorLandscape'
        belief_architecture:
          $ref: '#/components/schemas/BeliefArchitecture'
        objections_section:
          $ref: '#/components/schemas/ObjectionsSection'
        research_inspiration:
          $ref: '#/components/schemas/ResearchInspiration'
        # Legacy fields
        potential_product_names:
          type: array
          items:
            type: string
        level_of_consciousness:
          $ref: '#/components/schemas/ConsciousnessLevel'
        level_of_awareness:
          $ref: '#/components/schemas/AwarenessLevel'
        stage_of_sophistication:
          $ref: '#/components/schemas/StageOfSophistication'
        big_idea:
          type: string
          nullable: true
        metaphors:
          type: array
          items:
            type: string
        potential_ump:
          type: array
          items:
            type: string
        potential_ums:
          type: array
          items:
            type: string
        guru:
          type: string
          nullable: true
        discovery_story:
          type: string
          nullable: true
        headline_ideas:
          type: array
          items:
            $ref: '#/components/schemas/HeadlineIdea'
        objections:
          type: array
          items:
            type: string
        belief_chains:
          type: array
          items:
            $ref: '#/components/schemas/BeliefChain'
        funnel_architecture:
          type: array
          items:
            type: string
        potential_domains:
          type: array
          items:
            type: string
        examples_swipes:
          type: array
          items:
            $ref: '#/components/schemas/SwipeExample'
...
  headers:
    AccessControlAllowOrigin:
      schema:
        type: string
      example: '*'
    AccessControlAllowHeaders:
      schema:
        type: string
      example: 'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'
    AccessControlAllowMethods:
      schema:
        type: string
      example: 'GET,POST,OPTIONS'